// flutter/js/src/proto/message.ts
export interface VideoFrame {
  vp9s?: EncodedVideoFrames | undefined;
  rgb?: RGB | undefined;
  yuv?: YUV | undefined;
  h264s?: EncodedVideoFrames | undefined;
  h265s?: EncodedVideoFrames | undefined;
  vp8s?: EncodedVideoFrames | undefined;
  av1s?: EncodedVideoFrames | undefined;
  display: number;
}

export interface IdPk {
  id: string;
  pk: Uint8Array;
}
// ...
export interface LoginRequest {
  username: string;
  password: Uint8Array;
  my_id: string;
  my_name: string;
  option: OptionMessage | undefined;
  file_transfer?: FileTransfer | undefined;
  port_forward?: PortForward | undefined;
  video_ack_required: boolean;
  session_id: number;
  version: string;
  os_login: OSLogin | undefined;
  my_platform: string;
  hwid: Uint8Array;
}
// ...
export interface LoginResponse {
  error?: string | undefined;
  peer_info?: PeerInfo | undefined;
  enable_trusted_devices: boolean;
}
// ...
export interface PointerDeviceEvent {
  touch_event?: TouchEvent | undefined;
  modifiers: ControlKey[];
}

export interface MouseEvent {
  mask: number;
  x: number;
  y: number;
  modifiers: ControlKey[];
}
// ...
export interface CursorData {
  id: number;
  hotx: number;
  hoty: number;
  width: number;
  height: number;
  colors: Uint8Array;
}

export interface CursorPosition {
  x: number;
  y: number;
}

export interface Hash {
  salt: string;
  challenge: string;
}

export interface Clipboard {
  compress: boolean;
  content: Uint8Array;
  width: number;
  height: number;
  format: ClipboardFormat;
  /** Special format name, only used when format is Special. */
  special_name: string;
}
// ...
export interface TestDelay {
  time: number;
  from_client: boolean;
  last_delay: number;
  target_bitrate: number;
}

export interface PublicKey {
  asymmetric_value: Uint8Array;
  symmetric_value: Uint8Array;
}

export interface SignedId {
  id: Uint8Array;
}

export interface AudioFormat {
  sample_rate: number;
  channels: number;
}

export interface AudioFrame {
  data: Uint8Array;
}
// ...
export interface SwitchSidesResponse {
  uuid: Uint8Array;
  lr: LoginRequest | undefined;
}
// ...
export interface Message {
  signed_id?: SignedId | undefined;
  public_key?: PublicKey | undefined;
  test_delay?: TestDelay | undefined;
  video_frame?: VideoFrame | undefined;
  login_request?: LoginRequest | undefined;
  login_response?: LoginResponse | undefined;
  hash?: Hash | undefined;
  mouse_event?: MouseEvent | undefined;
  audio_frame?: AudioFrame | undefined;
  cursor_data?: CursorData | undefined;
  cursor_position?: CursorPosition | undefined;
  cursor_id?: number | undefined;
  key_event?: KeyEvent | undefined;
  clipboard?: Clipboard | undefined;
  file_action?: FileAction | undefined;
  file_response?: FileResponse | undefined;
  misc?: Misc | undefined;
  cliprdr?: Cliprdr | undefined;
  message_box?: MessageBox | undefined;
  switch_sides_response?: SwitchSidesResponse | undefined;
  voice_call_request?: VoiceCallRequest | undefined;
  voice_call_response?: VoiceCallResponse | undefined;
  peer_info?: PeerInfo | undefined;
  pointer_device_event?: PointerDeviceEvent | undefined;
  auth_2fa?: Auth2FA | undefined;
  multi_clipboards?: MultiClipboards | undefined;
}
// ...
export const IdPk = {
  encode(message: IdPk, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.id !== "") {
      writer.uint32(10).string(message.id);
    }
    if (message.pk.length !== 0) {
      writer.uint32(18).bytes(message.pk);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): IdPk {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseIdPk();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.id = reader.string();
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          message.pk = reader.bytes();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): IdPk {
    return {
      id: isSet(object.id) ? globalThis.String(object.id) : "",
      pk: isSet(object.pk) ? bytesFromBase64(object.pk) : new Uint8Array(0),
    };
  },

  toJSON(message: IdPk): unknown {
    const obj: any = {};
    if (message.id !== "") {
      obj.id = message.id;
    }
    if (message.pk.length !== 0) {
      obj.pk = base64FromBytes(message.pk);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<IdPk>, I>>(base?: I): IdPk {
    return IdPk.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<IdPk>, I>>(object: I): IdPk {
    const message = createBaseIdPk();
    message.id = object.id ?? "";
    message.pk = object.pk ?? new Uint8Array(0);
    return message;
  },
};
// ...
export const TestDelay = {
  encode(message: TestDelay, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.time !== 0) {
      writer.uint32(8).int64(message.time);
    }
    if (message.from_client !== false) {
      writer.uint32(16).bool(message.from_client);
    }
    if (message.last_delay !== 0) {
      writer.uint32(24).uint32(message.last_delay);
    }
    if (message.target_bitrate !== 0) {
      writer.uint32(32).uint32(message.target_bitrate);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): TestDelay {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseTestDelay();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.time = longToNumber(reader.int64() as Long);
          continue;
        case 2:
          if (tag !== 16) {
            break;
          }

          message.from_client = reader.bool();
          continue;
        case 3:
          if (tag !== 24) {
            break;
          }

          message.last_delay = reader.uint32();
          continue;
        case 4:
          if (tag !== 32) {
            break;
          }

          message.target_bitrate = reader.uint32();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): TestDelay {
    return {
      time: isSet(object.time) ? globalThis.Number(object.time) : 0,
      from_client: isSet(object.from_client) ? globalThis.Boolean(object.from_client) : false,
      last_delay: isSet(object.last_delay) ? globalThis.Number(object.last_delay) : 0,
      target_bitrate: isSet(object.target_bitrate) ? globalThis.Number(object.target_bitrate) : 0,
    };
  },

  toJSON(message: TestDelay): unknown {
    const obj: any = {};
    if (message.time !== 0) {
      obj.time = Math.round(message.time);
    }
    if (message.from_client !== false) {
      obj.from_client = message.from_client;
    }
    if (message.last_delay !== 0) {
      obj.last_delay = Math.round(message.last_delay);
    }
    if (message.target_bitrate !== 0) {
      obj.target_bitrate = Math.round(message.target_bitrate);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<TestDelay>, I>>(base?: I): TestDelay {
    return TestDelay.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<TestDelay>, I>>(object: I): TestDelay {
    const message = createBaseTestDelay();
    message.time = object.time ?? 0;
    message.from_client = object.from_client ?? false;
    message.last_delay = object.last_delay ?? 0;
    message.target_bitrate = object.target_bitrate ?? 0;
    return message;
  },
};

function createBasePublicKey(): PublicKey {
  return { asymmetric_value: new Uint8Array(0), symmetric_value: new Uint8Array(0) };
}

export const PublicKey = {
  encode(message: PublicKey, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.asymmetric_value.length !== 0) {
      writer.uint32(10).bytes(message.asymmetric_value);
    }
    if (message.symmetric_value.length !== 0) {
      writer.uint32(18).bytes(message.symmetric_value);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): PublicKey {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasePublicKey();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.asymmetric_value = reader.bytes();
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          message.symmetric_value = reader.bytes();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): PublicKey {
    return {
      asymmetric_value: isSet(object.asymmetric_value) ? bytesFromBase64(object.asymmetric_value) : new Uint8Array(0),
      symmetric_value: isSet(object.symmetric_value) ? bytesFromBase64(object.symmetric_value) : new Uint8Array(0),
    };
  },

  toJSON(message: PublicKey): unknown {
    const obj: any = {};
    if (message.asymmetric_value.length !== 0) {
      obj.asymmetric_value = base64FromBytes(message.asymmetric_value);
    }
    if (message.symmetric_value.length !== 0) {
      obj.symmetric_value = base64FromBytes(message.symmetric_value);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<PublicKey>, I>>(base?: I): PublicKey {
    return PublicKey.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<PublicKey>, I>>(object: I): PublicKey {
    const message = createBasePublicKey();
    message.asymmetric_value = object.asymmetric_value ?? new Uint8Array(0);
    message.symmetric_value = object.symmetric_value ?? new Uint8Array(0);
    return message;
  },
};

function createBaseSignedId(): SignedId {
  return { id: new Uint8Array(0) };
}

export const SignedId = {
  encode(message: SignedId, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.id.length !== 0) {
      writer.uint32(10).bytes(message.id);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): SignedId {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSignedId();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.id = reader.bytes();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SignedId {
    return { id: isSet(object.id) ? bytesFromBase64(object.id) : new Uint8Array(0) };
  },

  toJSON(message: SignedId): unknown {
    const obj: any = {};
    if (message.id.length !== 0) {
      obj.id = base64FromBytes(message.id);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<SignedId>, I>>(base?: I): SignedId {
    return SignedId.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SignedId>, I>>(object: I): SignedId {
    const message = createBaseSignedId();
    message.id = object.id ?? new Uint8Array(0);
    return message;
  },
};
// ...
export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };