// flutter/js/src/proto/rendezvous.ts
export enum ConnType {
  DEFAULT_CONN = 0,
  FILE_TRANSFER = 1,
  PORT_FORWARD = 2,
  RDP = 3,
  UNRECOGNIZED = -1,
}
// ...
export enum NatType {
  UNKNOWN_NAT = 0,
  ASYMMETRIC = 1,
  SYMMETRIC = 2,
  UNRECOGNIZED = -1,
}
// ...
export interface RegisterPeer {
  id: string;
  serial: number;
}

export interface RegisterPeerResponse {
  request_pk: boolean;
}

export interface PunchHoleRequest {
  id: string;
  nat_type: NatType;
  licence_key: string;
  conn_type: ConnType;
  token: string;
  version: string;
}
// ...
export enum PunchHoleResponse_Failure {
  ID_NOT_EXIST = 0,
  OFFLINE = 2,
  LICENSE_MISMATCH = 3,
  LICENSE_OVERUSE = 4,
  UNRECOGNIZED = -1,
}
// ...
export interface RequestRelay {
  id: string;
  uuid: string;
  socket_addr: Uint8Array;
  relay_server: string;
  secure: boolean;
  licence_key: string;
  conn_type: ConnType;
  token: string;
}

export interface RelayResponse {
  socket_addr: Uint8Array;
  uuid: string;
  relay_server: string;
  id?: string | undefined;
  pk?: Uint8Array | undefined;
  refuse_reason: string;
  version: string;
  feedback: number;
}

export interface SoftwareUpdate {
  url: string;
}

/**
 * if in same intranet, punch hole won't work both for udp and tcp,
 * even some router has below connection error if we connect itself,
 *  { kind: Other, error: "could not resolve to any address" },
 * so we request local address to connect.
 */
export interface FetchLocalAddr {
  socket_addr: Uint8Array;
  relay_server: string;
}
// ...
export interface RendezvousMessage {
  register_peer?: RegisterPeer | undefined;
  register_peer_response?: RegisterPeerResponse | undefined;
  punch_hole_request?: PunchHoleRequest | undefined;
  punch_hole?: PunchHole | undefined;
  punch_hole_sent?: PunchHoleSent | undefined;
  punch_hole_response?: PunchHoleResponse | undefined;
  fetch_local_addr?: FetchLocalAddr | undefined;
  local_addr?: LocalAddr | undefined;
  configure_update?: ConfigUpdate | undefined;
  register_pk?: RegisterPk | undefined;
  register_pk_response?: RegisterPkResponse | undefined;
  software_update?: SoftwareUpdate | undefined;
  request_relay?: RequestRelay | undefined;
  relay_response?: RelayResponse | undefined;
  test_nat_request?: TestNatRequest | undefined;
  test_nat_response?: TestNatResponse | undefined;
  peer_discovery?: PeerDiscovery | undefined;
  online_request?: OnlineRequest | undefined;
  online_response?: OnlineResponse | undefined;
  key_exchange?: KeyExchange | undefined;
  hc?: HealthCheck | undefined;
}
// ...
export const RequestRelay = {
  encode(message: RequestRelay, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.id !== "") {
      writer.uint32(10).string(message.id);
    }
    if (message.uuid !== "") {
      writer.uint32(18).string(message.uuid);
    }
    if (message.socket_addr.length !== 0) {
      writer.uint32(26).bytes(message.socket_addr);
    }
    if (message.relay_server !== "") {
      writer.uint32(34).string(message.relay_server);
    }
    if (message.secure !== false) {
      writer.uint32(40).bool(message.secure);
    }
    if (message.licence_key !== "") {
      writer.uint32(50).string(message.licence_key);
    }
    if (message.conn_type !== 0) {
      writer.uint32(56).int32(message.conn_type);
    }
    if (message.token !== "") {
      writer.uint32(66).string(message.token);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): RequestRelay {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseRequestRelay();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.id = reader.string();
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          message.uuid = reader.string();
          continue;
        case 3:
          if (tag !== 26) {
            break;
          }

          message.socket_addr = reader.bytes();
          continue;
        case 4:
          if (tag !== 34) {
            break;
          }

          message.relay_server = reader.string();
          continue;
        case 5:
          if (tag !== 40) {
            break;
          }

          message.secure = reader.bool();
          continue;
        case 6:
          if (tag !== 50) {
            break;
          }

          message.licence_key = reader.string();
          continue;
        case 7:
          if (tag !== 56) {
            break;
          }

          message.conn_type = reader.int32() as any;
          continue;
        case 8:
          if (tag !== 66) {
            break;
          }

          message.token = reader.string();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): RequestRelay {
    return {
      id: isSet(object.id) ? globalThis.String(object.id) : "",
      uuid: isSet(object.uuid) ? globalThis.String(object.uuid) : "",
      socket_addr: isSet(object.socket_addr) ? bytesFromBase64(object.socket_addr) : new Uint8Array(0),
      relay_server: isSet(object.relay_server) ? globalThis.String(object.relay_server) : "",
      secure: isSet(object.secure) ? globalThis.Boolean(object.secure) : false,
      licence_key: isSet(object.licence_key) ? globalThis.String(object.licence_key) : "",
      conn_type: isSet(object.conn_type) ? connTypeFromJSON(object.conn_type) : 0,
      token: isSet(object.token) ? globalThis.String(object.token) : "",
    };
  },

  toJSON(message: RequestRelay): unknown {
    const obj: any = {};
    if (message.id !== "") {
      obj.id = message.id;
    }
    if (message.uuid !== "") {
      obj.uuid = message.uuid;
    }
    if (message.socket_addr.length !== 0) {
      obj.socket_addr = base64FromBytes(message.socket_addr);
    }
    if (message.relay_server !== "") {
      obj.relay_server = message.relay_server;
    }
    if (message.secure !== false) {
      obj.secure = message.secure;
    }
    if (message.licence_key !== "") {
      obj.licence_key = message.licence_key;
    }
    if (message.conn_type !== 0) {
      obj.conn_type = connTypeToJSON(message.conn_type);
    }
    if (message.token !== "") {
      obj.token = message.token;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<RequestRelay>, I>>(base?: I): RequestRelay {
    return RequestRelay.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<RequestRelay>, I>>(object: I): RequestRelay {
    const message = createBaseRequestRelay();
    message.id = object.id ?? "";
    message.uuid = object.uuid ?? "";
    message.socket_addr = object.socket_addr ?? new Uint8Array(0);
    message.relay_server = object.relay_server ?? "";
    message.secure = object.secure ?? false;
    message.licence_key = object.licence_key ?? "";
    message.conn_type = object.conn_type ?? 0;
    message.token = object.token ?? "";
    return message;
  },
};

function createBaseRelayResponse(): RelayResponse {
  return {
    socket_addr: new Uint8Array(0),
    uuid: "",
    relay_server: "",
    id: undefined,
    pk: undefined,
    refuse_reason: "",
    version: "",
    feedback: 0,
  };
}

export const RelayResponse = {
  encode(message: RelayResponse, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.socket_addr.length !== 0) {
      writer.uint32(10).bytes(message.socket_addr);
    }
    if (message.uuid !== "") {
      writer.uint32(18).string(message.uuid);
    }
    if (message.relay_server !== "") {
      writer.uint32(26).string(message.relay_server);
    }
    if (message.id !== undefined) {
      writer.uint32(34).string(message.id);
    }
    if (message.pk !== undefined) {
      writer.uint32(42).bytes(message.pk);
    }
    if (message.refuse_reason !== "") {
      writer.uint32(50).string(message.refuse_reason);
    }
    if (message.version !== "") {
      writer.uint32(58).string(message.version);
    }
    if (message.feedback !== 0) {
      writer.uint32(72).int32(message.feedback);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): RelayResponse {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseRelayResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.socket_addr = reader.bytes();
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          message.uuid = reader.string();
          continue;
        case 3:
          if (tag !== 26) {
            break;
          }

          message.relay_server = reader.string();
          continue;
        case 4:
          if (tag !== 34) {
            break;
          }

          message.id = reader.string();
          continue;
        case 5:
          if (tag !== 42) {
            break;
          }

          message.pk = reader.bytes();
          continue;
        case 6:
          if (tag !== 50) {
            break;
          }

          message.refuse_reason = reader.string();
          continue;
        case 7:
          if (tag !== 58) {
            break;
          }

          message.version = reader.string();
          continue;
        case 9:
          if (tag !== 72) {
            break;
          }

          message.feedback = reader.int32();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): RelayResponse {
    return {
      socket_addr: isSet(object.socket_addr) ? bytesFromBase64(object.socket_addr) : new Uint8Array(0),
      uuid: isSet(object.uuid) ? globalThis.String(object.uuid) : "",
      relay_server: isSet(object.relay_server) ? globalThis.String(object.relay_server) : "",
      id: isSet(object.id) ? globalThis.String(object.id) : undefined,
      pk: isSet(object.pk) ? bytesFromBase64(object.pk) : undefined,
      refuse_reason: isSet(object.refuse_reason) ? globalThis.String(object.refuse_reason) : "",
      version: isSet(object.version) ? globalThis.String(object.version) : "",
      feedback: isSet(object.feedback) ? globalThis.Number(object.feedback) : 0,
    };
  },

  toJSON(message: RelayResponse): unknown {
    const obj: any = {};
    if (message.socket_addr.length !== 0) {
      obj.socket_addr = base64FromBytes(message.socket_addr);
    }
    if (message.uuid !== "") {
      obj.uuid = message.uuid;
    }
    if (message.relay_server !== "") {
      obj.relay_server = message.relay_server;
    }
    if (message.id !== undefined) {
      obj.id = message.id;
    }
    if (message.pk !== undefined) {
      obj.pk = base64FromBytes(message.pk);
    }
    if (message.refuse_reason !== "") {
      obj.refuse_reason = message.refuse_reason;
    }
    if (message.version !== "") {
      obj.version = message.version;
    }
    if (message.feedback !== 0) {
      obj.feedback = Math.round(message.feedback);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<RelayResponse>, I>>(base?: I): RelayResponse {
    return RelayResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<RelayResponse>, I>>(object: I): RelayResponse {
    const message = createBaseRelayResponse();
    message.socket_addr = object.socket_addr ?? new Uint8Array(0);
    message.uuid = object.uuid ?? "";
    message.relay_server = object.relay_server ?? "";
    message.id = object.id ?? undefined;
    message.pk = object.pk ?? undefined;
    message.refuse_reason = object.refuse_reason ?? "";
    message.version = object.version ?? "";
    message.feedback = object.feedback ?? 0;
    return message;
  },
};
// ...
export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };